user-service/requirements.txt
# User Service - Independent Dependencies (SQLAlchemy + Pydantic)
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database (SQLAlchemy instead of SQLModel for stability)
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# HTTP Client (for inter-service communication)
httpx==0.27.0

# Utils
python-dotenv==1.0.1
email-validator==2.1.0
loguru==0.7.2

# Development & Testing
pytest==8.3.0
pytest-asyncio==0.23.0